.windsurf/skills/dtx/dtx-gather/scripts/include/pyproject.toml
[project]
name = "dtx-expand-cli"
version = "0.1.0"
description = "CLI for dtx-expand skill"
requires-python = ">=3.10"
dependencies = [
    "pyyaml>=6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
